+++ core/src/main/scala/ai/lum/odinson/lucene/search/OdinsonCollector.scala
@@ -11,18 +11,19 @@ class OdinsonCollector(
   private val collectedResults: Array[OdinsonScoreDoc],
   private val after: Int,
   private val computeTotalHits: Boolean,
+  private val disableMatchSelector: Boolean,
 ) extends Collector {
 
-  def this(numHits: Int, after: Int, computeTotalHits: Boolean) = {
-    this(new Array[OdinsonScoreDoc](numHits), after, computeTotalHits)
+  def this(numHits: Int, after: Int, computeTotalHits: Boolean, disableMatchSelector: Boolean) = {
+    this(new Array[OdinsonScoreDoc](numHits), after, computeTotalHits, disableMatchSelector)
   }
 
-  def this(numHits: Int, computeTotalHits: Boolean) = {
-    this(numHits, -1, computeTotalHits)
+  def this(numHits: Int, computeTotalHits: Boolean, disableMatchSelector: Boolean) = {
+    this(numHits, -1, computeTotalHits, disableMatchSelector)
   }
 
-  def this(numHits: Int, afterDoc: OdinsonScoreDoc, computeTotalHits: Boolean) = {
-    this(numHits, if (afterDoc == null) -1 else afterDoc.doc, computeTotalHits)
+  def this(numHits: Int, afterDoc: OdinsonScoreDoc, computeTotalHits: Boolean, disableMatchSelector: Boolean) = {
+    this(numHits, if (afterDoc == null) -1 else afterDoc.doc, computeTotalHits, disableMatchSelector)
   }
 
   private var totalHits: Int = 0
@@ -52,17 +53,23 @@ class OdinsonCollector(
     }
   }
 
-  case class TotalHitsCalculatingLeafCollector(docBase: Int, afterDoc: Int) extends OdinsonLeafCollector {
+  case class TotalHitsCalculatingLeafCollector(
+    docBase: Int,
+    afterDoc: Int,
+  ) extends OdinsonLeafCollector {
     def collect(doc: Int): Unit = {
       totalHits += 1
       if (collectedHits >= collectedResults.length || doc <= afterDoc) {
-        return // don't terminate, we want to keep collecting for accurate totalHits count
+        // don't terminate, we want to keep collecting
+        // for accurate totalHits count
+        return
       }
+      val matches = if (disableMatchSelector) scorer.getAllPossibleMatches() else scorer.getMatches()
       collectedResults(collectedHits) = new OdinsonScoreDoc(
         doc = doc + docBase,
         score = scorer.score(),
         shardIndex = -1,
-        matches = scorer.getMatches(),
+        matches = matches,
         segmentDocId = doc,
         segmentDocBase = docBase,
       )
@@ -70,19 +77,24 @@ class OdinsonCollector(
     }
   }
 
-  case class EarlyTerminationLeafCollector(docBase: Int, afterDoc: Int) extends OdinsonLeafCollector {
+  case class EarlyTerminationLeafCollector(
+    docBase: Int,
+    afterDoc: Int,
+  ) extends OdinsonLeafCollector {
     def collect(doc: Int): Unit = {
       if (collectedHits >= collectedResults.length) {
-        throw new CollectionTerminatedException() // terminate, since all required results have been collected
+        // terminate, since all required results have been collected
+        throw new CollectionTerminatedException()
       }
       if (doc <= afterDoc) {
         return
       }
+      val matches = if (disableMatchSelector) scorer.getAllPossibleMatches() else scorer.getMatches()
       collectedResults(collectedHits) = new OdinsonScoreDoc(
         doc = doc + docBase,
         score = scorer.score(),
         shardIndex = -1,
-        matches = scorer.getMatches(),
+        matches = matches,
         segmentDocId = doc,
         segmentDocBase = docBase,
       )