+++ core/src/test/scala/ai/lum/odinson/TestPatterns.scala
@@ -15,7 +15,8 @@ class TestPatterns extends FlatSpec with Matchers {
     val expected = allExpected.split(";", -1)
     pattern should s"find all expected results for ${string.display}" in {
       val ee = TestUtils.mkExtractorEngine(string)
-      val results = ee.query(pattern)
+      val q = ee.compiler.mkQuery(pattern)
+      val results = ee.query(q)
       val actual = TestUtils.mkStrings(results, ee)
       actual should equal (expected)
     }