+++ extra/src/main/scala/ai/lum/odinson/extra/Shell.scala
@@ -149,10 +149,6 @@ object Shell extends App {
     println()
     println(s"Scala version: ${BuildInfo.scalaVersion}")
     println(s"Sbt version: ${BuildInfo.sbtVersion}")
-    println(s"Dependencies:")
-    BuildInfo.libraryDependencies.foreach(s => println("  " + s))
-    println(s"Scalac options:")
-    BuildInfo.scalacOptions.foreach(s => println("  " + s))
     println()
   }
 
@@ -163,7 +159,8 @@ object Shell extends App {
   /** searches for pattern and prints the first n matches */
   def search(n: Int): Unit = {
     val start = System.currentTimeMillis()
-    val results = extractorEngine.query(query, n)
+    val q = extractorEngine.compiler.mkQuery(query)
+    val results = extractorEngine.query(q, n)
     val duration = (System.currentTimeMillis() - start) / 1000f
     after = results.scoreDocs.lastOption.getOrElse(null)
     totalHits = results.totalHits
@@ -182,7 +179,8 @@ object Shell extends App {
       return
     }
     val start = System.currentTimeMillis()
-    val results = extractorEngine.query(query, n, after)
+    val q = extractorEngine.compiler.mkQuery(query)
+    val results = extractorEngine.query(q, n, after)
     val duration = (System.currentTimeMillis() - start) / 1000f
     after = results.scoreDocs.lastOption.getOrElse(null)
     if (after == null) {